+++ lib/interactive_svg_widget.dart
@@ -1,7 +1,9 @@
 library interactive_svg_widget;
 
-// WIDGET
-export 'src/widgets/interactive_svg_widget.dart';
 
 // MODELS
+export 'src/models/gesture_detector_data.dart';
+export 'src/models/interactive_svg_theme_data.dart';
 export 'src/models/region.dart';
+// WIDGET
+export 'src/widgets/interactive_svg_widget.dart';